Move _client into _utils
